sprint-6/step-2 renamed tests
